Use the system atof() on vax instead of perl's
Because numbers on vax are special.

Thanks to miod@ for noting that they are special and leont on
irc.perl.org#p5p for the fix.

This resolves some Floating point exceptions in perl, although there
are still some things that vax gets wrong.

Probably should figure out how to make this the default for vax though.