Longer timeout for COW test
Because on my alpha it fails after just a couple of seconds,
not the tens of seconds that would indicate an actual problem.

It would be better if this test calculated how fast the machine
it is running on is and used that to figure out how fast things
should happen, but that's more work than just massaging the numbers.