Successfull build logs for my suite of test machines
Although the alpha test was not successful, it was because my test machine
is too slow and it was a speed test.  I increased the timeout, but really
the timeouts need a "factor" calculated at the beginning and then used to
adjust the actual timeouts to make it more accurate.