Fixes for CVE
Date: Mon, 18 May 2020 12:25:40 -0500
From: John Lightsey <john@nixnuts.net>
Subject: Notification of upcoming Perl security release

* CVE-2020-10543

Found by: ManhND of The Tarantula Team, VinCSS (a member of Vingroup)
Fixed by: John Lightsey
Versions affected: 5.005 to 5.30.2

This is a heap buffer overflow in Perl's regular expression compiler
that overwrites memory allocated after the regular expression storage
space with attacker supplied data. The heap overflow occurs due to a
signed size_t integer overflow in the storage space calculations for
nested regular expression quantifiers.

An application written in Perl would only be vulnerable to this flaw if
it evaluates regular expressions supplied by the attacker. Evaluating
regular expressions in this fashion is known to be dangerous since the
regular expression engine does not protect against denial of service
attacks in this usage scenario.

Additionally, the target system needs a sufficient amount of memory to
allocate partial expansions of the nested quantifiers prior to the
overflow occurring. This requirement is unlikely to be met on 64bit
systems.

Patches for the supported Perl versions are attached to this email.

5.30: v530-0001-CVE-2020-10543.patch

5.28: v528-0001-CVE-2020-10543.patch

* CVE-2020-10878

Found by: Hugo van der Sanden and Slaven Rezic
Fixed by: Hugo van der Sanden and Karl Williamson
Versions affected: 5.005 to 5.30.2

This vulnerability is an attacker controlled corruption of the
intermediate language state of a compiled regular expression. The
corruption occurs due to integer overflows in the calculation of offsets
between instructions for the regular expression engine. An attacker
could abuse this behavior to insert instructions into the compiled form
of a Perl regular expression.

An application written in Perl would only be vulnerable to this flaw if
it evaluates regular expressions supplied by the attacker. Evaluating
regular expressions in this fashion is known to be dangerous since the
regular expression engine does not protect against denial of service
attacks in this usage scenario.

Patches for the supported Perl versions are attached to this email.

5.30: v530-0002-CVE-2020-10878.patch
      v530-0003-CVE-2020-10878.patch

5.28: v528-0002-CVE-2020-10878.patch

* CVE-2020-12723

Found by: Sergey Aleynikov
Fixed by: Hugo van der Sanden
Versions affected: 5.10.0 to 5.30.2

This vulnerability is an attacker controlled corruption of the
intermediate language state of a compiled regular expression. The
corruption occurs when Perl's regular expression compiler calls
S_study_chunk() in a recursive fashion to optimize the intermediate
language representation of a regular expression. An attacker could abuse
this behavior to insert instructions into the compiled form of a Perl
regular expression.

An application written in Perl would only be vulnerable to this flaw if
it evaluates regular expressions supplied by the attacker. Evaluating
regular expressions in this fashion is known to be dangerous since the
regular expression engine does not protect against denial of service
attacks in this usage scenario.

Patches for the supported Perl versions are attached to this email.

5.30: v530-0004-CVE-2020-12723.patch

5.28: v528-0003-CVE-2020-12723.patch