Add build logs from the rest of my machines
Well, apparently I don't have an i386 at the moment, will have to
figure out where I put that.

The alpha build did fail some tests, but I think that's because it's slow as
the current perl version fails that test too.  I don't recall that
it did before but it does now.