Add luna88k build log
From Kenji Aoyama <aoyama () nk-home ! net>

This fails a bunch of speed tests, unsurprisingly as it didn't use the
build infrastructure that set the timeout multipler, but it also fails
some Inf/NaN tests which is a bit weird.  It also fails the
OpenBSD::Unveil -w test, which ddickman@ ran into as well.

Very strange.