Unpack arguments into variables first.
miod@ pointed out that function argument order is not always left
to right, and in some cases it is definitely right to left so we
have to use variables first to make sure we unpack the vargs in the
right order.

deraadt@ gave me the clue about the extra block to hide
the variables.