Pick up patch to skip permission tests
As the patch says, cvs doesn't store the permissions and so things
are incorrect and the tests don't pass.  We fix the permissions
during install so it doesn't matter except that the tests don't
pass.  Since bluhm@ is going to start running the perl tests
regularly, we want them to pass.