Use readline instead of getline
In this special case of `while (readline $fh)` perl expands this to
`while (defined $_ = readline $fh)` which is a more correct check.

The failure would be an invalid text file without a trailing newline
where the only thing on the line is "0", which would not be valid
in any of the files we are parsing, but correct is better than "works".

Suggested by gkoehler@