Unpack va_args as long and cast to proper type
In pp_sys.c, perl doesn't try very hard and every argument to syscall
is "long", because we know that, we can do the translation in the
emulator and instead of having to mess around with faking things,
it (mostly) Just Works.

On macppc where sizes differ:

Before:
$ perl 3</etc/passwd -E  'syscall(169, 3, $_ = "c" x 7, 7, 0, 0, 11); say'
Charlie

After:
$ perl 3</etc/passwd -E  'syscall(169, 3, $_ = "c" x 7, 7, 11); say'
Charlie

On sparc64 and amd64, it's the second option before and after.

Example and suggestion from gkehler@